Feature: Main Product Section
Created the `main-product` section to showcase details of a single
product. Added multiple blocks for the page to be customized.

1. The product media is shown as a static gallery with thumbnails and
the user has options to place the thumbnails in 3 different positions.

2. Other blocks have also been added to display product details like
description, SKU, Vendor name, Price.

3. Added a block to show the product variants as buttons. When a
variant option is chosen, the value of the input[name="id"] is set in
the product form.

4. Added a product form block, which contains a quantity selector,
which the user can disable as well.